Warn user and prevent save if new page title conflicts with existing page URL
