+++ conftest.py
@@ -10,7 +10,7 @@ from django.db.models.signals import pre_save
 from fixtures.common import *  # pylint:disable=wildcard-import,unused-wildcard-import  # noqa: F403
 from websites.constants import OMNIBUS_STARTER_SLUG
 from websites.models import WebsiteContent, WebsiteStarter
-from websites.signals import update_page_url_on_title_change
+from websites.signals import update_navmenu_on_title_change
 from websites.site_config_api import SiteConfig
 
 
@@ -96,7 +96,7 @@ def disable_websitecontent_signal():
     """Disable page url update signal"""
 
     # Disconnect
-    pre_save.disconnect(update_page_url_on_title_change, sender=WebsiteContent)
+    pre_save.disconnect(update_navmenu_on_title_change, sender=WebsiteContent)
 
 
 @pytest.fixture
@@ -104,4 +104,4 @@ def enable_websitecontent_signal():
     """Enable page url update signal"""
 
     # Disconnect
-    pre_save.connect(update_page_url_on_title_change, sender=WebsiteContent)
+    pre_save.connect(update_navmenu_on_title_change, sender=WebsiteContent)
